streamer: Add nonblocking quic server
