+++ streamer/src/nonblocking/mod.rs
@@ -1,2 +1,3 @@
+pub mod quic;
 pub mod recvmmsg;
 pub mod sendmmsg;
